Refactor attngan model code
