+++ picturate/models/attngan/__init__.py
@@ -1,7 +1,4 @@
 from picturate.models.attngan.attention import *
-from picturate.models.attngan.upsample import *
-from picturate.models.attngan.glu import *
-from picturate.models.attngan.conv import *
 from picturate.models.attngan.model import *
 
 from picturate.models.attngan.attention import AttentionGeneral as ATT_NET
