--- picturate/models/attngan/conv.py
@@ -1,53 +1,0 @@
-from picturate.imports import *
-from picturate.models.attngan import *
-
-
-def conv1x1(in_planes, out_planes, bias=False):
-    "1x1 convolution with padding"
-    return nn.Conv2d(
-        in_planes, out_planes, kernel_size=1, stride=1, padding=0, bias=bias
-    )
-
-
-def conv3x3(in_planes, out_planes):
-    "3x3 convolution with padding"
-    return nn.Conv2d(
-        in_planes, out_planes, kernel_size=3, stride=1, padding=1, bias=False
-    )
-
-
-# Upsale the spatial size by a factor of 2
-def upBlock(in_planes, out_planes):
-    block = nn.Sequential(
-        Upsample(scale_factor=2, mode="nearest"),
-        conv3x3(in_planes, out_planes * 2),
-        nn.BatchNorm2d(out_planes * 2),
-        GLU(),
-    )
-    return block
-
-
-# Keep the spatial size
-def Block3x3_relu(in_planes, out_planes):
-    block = nn.Sequential(
-        conv3x3(in_planes, out_planes * 2), nn.BatchNorm2d(out_planes * 2), GLU()
-    )
-    return block
-
-
-class ResBlock(nn.Module):
-    def __init__(self, channel_num):
-        super(ResBlock, self).__init__()
-        self.block = nn.Sequential(
-            conv3x3(channel_num, channel_num * 2),
-            nn.BatchNorm2d(channel_num * 2),
-            GLU(),
-            conv3x3(channel_num, channel_num),
-            nn.BatchNorm2d(channel_num),
-        )
-
-    def forward(self, x):
-        residual = x
-        out = self.block(x)
-        out += residual
-        return out
