--- picturate/models/attngan/upsample.py
@@ -1,26 +1,0 @@
-from picturate.imports import *
-from picturate.models.attngan import *
-
-
-class Upsample(nn.Module):
-    def __init__(
-        self, size=None, scale_factor=None, mode="nearest", align_corners=None
-    ):
-        super().__init__()
-        self.size = size
-        self.scale_factor = float(scale_factor) if scale_factor else None
-        self.mode = mode
-        self.align_corners = align_corners
-
-    def forward(self, x):
-        return F.interpolate(
-            x, self.size, self.scale_factor, self.mode, self.align_corners
-        )
-
-    def extra_repr(self):
-        if self.scale_factor is not None:
-            info = "scale_factor=" + str(self.scale_factor)
-        else:
-            info = "size=" + str(self.size)
-        info += ", mode=" + self.mode
-        return info
